+++ api/api/controllers/ciscat_controller.py
@@ -8,7 +8,7 @@ from typing import List
 
 import wazuh.ciscat as ciscat
 from wazuh.cluster.dapi.dapi import DistributedAPI
-from ..util import remove_nones_to_dict
+from ..util import remove_nones_to_dict, parse_search_param
 
 loop = asyncio.get_event_loop()
 logger = logging.getLogger('wazuh')
@@ -26,7 +26,8 @@ def get_agents_cistat_results(agent_id: str, pretty: bool = False, wait_for_comp
 
     :param agent_id: Agent ID. All posible values since 000 onwards.
     :param pretty: Show results in human-readable format 
-    :param wait_for_complete: Disable timeout response 
+    :param wait_for_complete: Disable timeout response
+    :param offset: First element to return in the collection
     :param limit: Maximum number of elements to return
     :param select: Select which fields to return (separated by comma)
     :param sort: Sorts the collection by a field or fields (separated by comma). Use +/- at the beginning to list in ascending or descending order. 
@@ -43,7 +44,7 @@ def get_agents_cistat_results(agent_id: str, pretty: bool = False, wait_for_comp
     f_kwargs = {'offset': offset,
                 'limit': limit,
                 'sort': sort,
-                'search': search,
+                'search': parse_search_param(search),
                 'select': select,
                 'agent_id': agent_id,
                 'filters': {