+++ api/api/util.py
@@ -149,3 +149,16 @@ def remove_nones_to_dict(dct):
     """
     return {k: v if not isinstance(v, dict) else remove_nones_to_dict(v)
             for k, v in dct.items() if v is not None}
+
+
+def parse_search_param(search: str):
+    """Parses search str param coming from the API query into a dictionary the framework can process.
+
+    :param search: Search parameter coming from the API query
+    :return: A dictionary like {'value': 'ubuntu', 'negation': False}
+    """
+    if search is not None:
+        negation = search[0] == '-'
+        return {'negation': negation, 'value': search[1:] if negation else search}
+    else:
+        return None